+++ bridge-adapters/kinetic-bridgehub-adapter-azure-rest/src/test/java/com/kineticdata/bridgehub/adapter/azure/rest/AzureRestTest.java
@@ -9,7 +9,6 @@ import com.kineticdata.bridgehub.adapter.RecordList;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import org.junit.Test;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;